requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
certifi==2021.10.8
click==8.1.3
fastapi==0.79.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
pydantic==1.9.1
python-dotenv==0.19.2
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
zipp==3.8.1